footer Controller移除，計時器換位置
