+++ app/views/rooms/_footer.html.erb
@@ -1,5 +1,4 @@
-<footer class="room-footer"
-        data-controller="footer">
+<footer class="room-footer">
   <div>
     <% if current_user %>
       <%= link_to "邀請", "/#{@room.uuid}/invite" %>
