+++ app/views/rooms/_webconsole.html.erb
@@ -12,7 +12,7 @@
       </button>
     </div>
     <div class="mr-6 flex items-center font-bold text-lg text-white"
-         id="timer">
+         data-allRoom-target="timer">
       00:00:01
     </div>
 </div>
